move Bazel stuff out of FlutterAppService
Also, removed TimeoutUtil since it's unused.
No user-visible changes.
TODO: move non-bazel command and remove FlutterAppService
